starting to add htmlcontroller - not functioning yet
