+++ app/views/create.handlebars
@@ -1,0 +1,26 @@
+<h2>Create an event:</h2>
+<hr>
+<form action="/api/event/new" method="POST">
+    <label for="userId">User ID</label>
+    <input type="text" name="userId">
+    <br>
+    <label for="eventName">Event Name</label>
+    <input type="text" name="eventName">
+    <br>
+    <label for="eventAddress">Event Address</label>
+    <input type="text" name="eventAddress">
+    <br>
+    <label for="eventCity">Event City</label>
+    <input type="text" name="eventCity">
+    <br>
+    <label for="eventState">Event State</label>
+    <input type="text" name="eventState">
+    <br>
+    <label for="eventZip">Event Zip</label>
+    <input type="text" name="eventZip">
+    <br>
+    <label for="eventCost">Cost</label>
+    <input type="text" name="eventCost">
+    <br>
+    <button type="submit">Submit</button>
+</form>
