+++ app/views/layouts/main.handlebars
@@ -6,7 +6,7 @@
   <!--Import Google Icon Font-->
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   <!--Import materialize.css-->
-    <link type="text/css" rel="stylesheet" href="./css/materialize/css/materialize.min.css" media="screen,projection"/>
+    <link type="text/css" rel="stylesheet" href="./css/materialize.min.css" media="screen,projection"/>
     <link rel="stylesheet" href="/public/css/test.css">
   <!--Let browser know website is optimized for mobile-->
     <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
@@ -18,16 +18,16 @@
       <div class="nav-wrapper">
         <a href="#!" class="brand-logo">itsaGO</a>
         <ul class="right hide-on-med-and-down">
-          <li><a href="login.html">Login</a></li>
-          <li><a href="dashboard.html">Event Dashboard</a></li>
-          <li><a href="about.html">About</a></li>
-          <li><a href="contact.html">Contact</a></li>
+          <li><a href="login">Login</a></li>
+          <li><a href="dashboard">Event Dashboard</a></li>
+          <li><a href="about">About</a></li>
+          <li><a href="contact">Contact</a></li>
         </ul>
         <ul class="side-nav" id="mobile-demo">
-          <li><a href="login.html">Login</a></li>
-          <li><a href="dashboard.html">Event Dashboard</a></li>
-          <li><a href="about.html">About</a></li>
-          <li><a href="contact.html">Contact</a></li>
+          <li><a href="login">Login</a></li>
+          <li><a href="dashboard">Event Dashboard</a></li>
+          <li><a href="about">About</a></li>
+          <li><a href="contact">Contact</a></li>
       </ul>
       </div>
     </nav>
@@ -46,7 +46,7 @@
 <!-- jQuery --> 
     <script type="text/javascript" src="https://code.jquery.com/jquery-3.2.1.min.js"></script>
 <!-- Materialize javascript -->
-    <script type="text/javascript" src="./css/materialize/js/materialize.min.js"></script>
+    <script type="text/javascript" src="./js/materialize.min.js"></script>
 
     <script type="text/javascript">
       $(document).ready(function(){
